feat: 도움 및 블로그 서비스에 관리자 권한 검사 추가
도움 및 블로그 서비스의 생성, 업데이트, 삭제 기능에 관리자 권한 검사를 추가하여 무단 접근을 방지합니다. 사용자의 그룹 정보를 확인하여 "/manager" 그룹에 속하지 않은 사용자는 403 Forbidden 오류를 반환합니다.

+++ src/app/blog/service/blog.py
@@ -57,6 +57,7 @@ class BlogService:
     ):
         if "/manager" not in user.groups:
             raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
         author = await self.author_repo.get_by_sub(session, user.sub)
         updated_author = {}
         if not author:
@@ -186,6 +187,9 @@ class BlogService:
         session: db_session,
         post_id: str = Path(),
     ):
+        if "/manager" not in user.groups:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
         # 포스트 불러오기
         post = await self.get_post_by_id(session, post_id)
         if not post:
@@ -237,4 +241,7 @@ class BlogService:
         session: db_session,
         post_id: str = Path(),
     ):
+        if "/manager" not in user.groups:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN)
+
         await self.blog_post_repo.delete(session, post_id)
